Fix: need to rebuild the table when growing it

+++ src/surgescript/util/fasthash.h
@@ -18,6 +18,7 @@
  * util/fasthash.h
  * A fast hash table with integer keys and linear probing
  */
+
 #ifndef _FASTHASH_H
 #define _FASTHASH_H
 
@@ -37,7 +38,7 @@ No need to compile fasthash.c separately */
 #endif
 
 typedef struct fasthash_t fasthash_t;
-FASTHASH_API fasthash_t* fasthash_create(void (*element_destructor)(void*), size_t lg2_cap);
+FASTHASH_API fasthash_t* fasthash_create(void (*element_destructor)(void*), int lg2_cap);
 FASTHASH_API fasthash_t* fasthash_destroy(fasthash_t* hashtable);
 FASTHASH_API void* fasthash_get(fasthash_t* hashtable, uint64_t key);
 FASTHASH_API void fasthash_put(fasthash_t* hashtable, uint64_t key, void* value);
